Bound derived matrix sizes in the C API
The shim checked M, P, and N*P, but the core stores matrix dimensions
in int and forms products of them (the largest matrix is the IPA
tableau of size (M+N) x (M+N+2)). A game with M = 65536 total actions
passed the checks and overflowed int in cmatrix's size computation,
which is undefined behaviour rather than an allocation failure. Require
M + N + 2 <= 46340, the largest value whose square fits in int, and
compute the two buffer sizes introduced earlier in size_t.

+++ cmatrix.cc
@@ -403,7 +403,7 @@ double cmatrix::adjoint() {
   double D = 1.0;
   // heap-allocated m-by-m work matrix (row-major); m*m doubles can exceed
   // the stack size
-  std::vector<double> retvalbuf(m * m);
+  std::vector<double> retvalbuf(static_cast<size_t>(m) * m);
   double *retval = retvalbuf.data();
 #define RETVAL(i,j) retval[(i)*m+(j)]
   for(i = 0; i < m; i++)
